# evidencias/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "device_systems API",
    "description": "API REST para la gestión de usuarios de device_systems. Utiliza FastAPI, Pydantic v2 y SQLAlchemy para persistencia relacional, CRUD completo, filtros, validaciones, constraints y manejo controlado de errores. Migraciones Alembic, dispositivos y préstamos.",
    "contact": {
      "name": "Nombre Apellido",
      "email": "ejemplo@device-systems.com"
    },
    "version": "4.0.0"
  },
  "paths": {
...
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Listar usuarios",
        "description": "Retorna todos los usuarios y permite filtrar por rol y estado.",
        "operationId": "list_users_users_get",
        "parameters": [
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/UserRole"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtra usuarios por rol: admin, support o user.",
              "title": "Role"
            },
            "description": "Filtra usuarios por rol: admin, support o user."
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtra usuarios por estado activo/inactivo.",
              "title": "Is Active"
            },
            "description": "Filtra usuarios por estado activo/inactivo."
          }
        ],
        "responses": {
          "200": {
            "description": "Listado de usuarios.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Registrar un nuevo usuario",
        "description": "Crea un usuario en la base de datos con validación Pydantic y restricciones SQLAlchemy.",
        "operationId": "create_user_users_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            }
          },
          "400": {
            "description": "Correo duplicado o restricción de base de datos"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Consultar usuario por ID",
        "description": "Retorna un usuario a partir de su ID.",
        "operationId": "get_user_users__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Datos públicos del usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Actualizar usuario",
        "description": "Reemplaza todos los campos del usuario existente.",
        "operationId": "replace_user_users__user_id__put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            }
          },
          "400": {
            "description": "Correo duplicado o restricción de base de datos"
          },
          "404": {
            "description": "Usuario no encontrado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Actualizar usuario parcialmente",
        "description": "Modifica solamente los campos enviados en el body.",
        "operationId": "update_user_users__user_id__patch",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            }
          },
          "400": {
            "description": "Sin campos, correo duplicado o restricción de base de datos"
          },
          "404": {
            "description": "Usuario no encontrado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Eliminar usuario",
        "description": "Elimina un usuario existente. Requiere la cabecera X-API-Key.",
        "operationId": "delete_user_users__user_id__delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Clave simulada de autenticación para operaciones sensibles.",
              "title": "X-Api-Key"
            },
            "description": "Clave simulada de autenticación para operaciones sensibles."
          }
        ],
        "responses": {
          "204": {
            "description": "Confirmación de eliminación."
          },
          "401": {
            "description": "API Key inválida o ausente"
          },
          "404": {
            "description": "Usuario no encontrado"
          },
          "409": {
            "description": "Usuario con historial de préstamos"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}/loans": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Préstamos del usuario",
        "description": "Consulta con joins el historial del usuario.",
        "operationId": "user_loans_users__user_id__loans_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Préstamos y equipos relacionados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LoanDetailResponse"
                  },
                  "title": "Response User Loans Users  User Id  Loans Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}/devices": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Dispositivos asignados",
        "description": "Equipos con préstamos active u overdue del usuario.",
        "operationId": "user_devices_users__user_id__devices_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dispositivos actualmente asignados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceResponse"
                  },
                  "title": "Response User Devices Users  User Id  Devices Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "serial_number": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Serial Number"
          },
          "device_type": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Device Type"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "is_available": {
            "type": "boolean",
            "title": "Is Available",
            "default": true
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "serial_number",
          "device_type",
          "id",
          "created_at"
        ],
        "title": "DeviceResponse",
        "example": {
          "brand": "lenovo",
          "device_type": "laptop",
          "name": "Lenovo ThinkPad",
          "serial_number": "LEN-001"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoanDetailResponse": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/LoanStatus"
          },
          "return_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return Date"
          },
          "user_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "User Id"
          },
          "device_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Device Id"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "loan_date": {
            "type": "string",
            "format": "date-time",
            "title": "Loan Date"
          },
          "user": {
            "$ref": "#/components/schemas/UserPublic"
          },
          "device": {
            "$ref": "#/components/schemas/DeviceResponse"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "return_date",
          "user_id",
          "device_id",
          "id",
          "loan_date",
          "user",
          "device"
        ],
        "title": "LoanDetailResponse",
        "example": {
          "device_id": 1,
          "user_id": 1
        }
      },
...
      "LoanStatus": {
        "type": "string",
        "enum": [
          "active",
          "returned",
          "overdue"
        ],
        "title": "LoanStatus"
      },
      "UserCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 3,
            "title": "Name",
            "description": "Nombre completo del usuario. Mínimo 3 caracteres.",
            "examples": [
              "Usuario Ejemplo"
            ]
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Correo electrónico único del usuario.",
            "examples": [
              "usuario@ejemplo.com"
            ]
          },
          "role": {
            "$ref": "#/components/schemas/UserRole",
            "description": "Rol del usuario dentro del sistema: admin, support o user.",
            "default": "user"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Indica si el usuario se encuentra activo en el sistema.",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "email"
        ],
        "title": "UserCreate",
        "description": "Esquema de entrada utilizado en POST /users.",
        "example": {
          "email": "usuario@ejemplo.com",
          "is_active": true,
          "name": "Usuario Ejemplo",
          "role": "admin"
        }
      },
      "UserListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Cantidad total de usuarios retornados."
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserPublic"
            },
            "type": "array",
            "title": "Items",
            "description": "Lista de usuarios."
          }
        },
        "type": "object",
        "required": [
          "total",
          "items"
        ],
        "title": "UserListResponse",
        "description": "Envoltorio estandarizado para la respuesta de listado de usuarios."
      },
      "UserPatch": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "UserPatch",
        "description": "Esquema de entrada utilizado en PATCH /users/{user_id}.\n\nTodos los campos son opcionales (default=None) porque el cliente solo\nenvía los que quiere modificar.",
        "example": {
          "role": "support"
        }
      },
      "UserPublic": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 3,
            "title": "Name",
            "description": "Nombre completo del usuario. Mínimo 3 caracteres.",
            "examples": [
              "Usuario Ejemplo"
            ]
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Correo electrónico único del usuario.",
            "examples": [
              "usuario@ejemplo.com"
            ]
          },
          "role": {
            "$ref": "#/components/schemas/UserRole",
            "description": "Rol del usuario dentro del sistema: admin, support o user.",
            "default": "user"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Indica si el usuario se encuentra activo en el sistema.",
            "default": true
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "email",
          "id"
        ],
        "title": "UserPublic",
        "description": "Response model utilizado por los endpoints de la API pública.",
        "example": {
          "email": "usuario@ejemplo.com",
          "id": 1,
          "is_active": true,
          "name": "Usuario Ejemplo",
          "role": "admin"
        }
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "admin",
          "support",
          "user"
        ],
        "title": "UserRole",
        "description": "Roles permitidos para un usuario del sistema device_systems."
      },
      "UserUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 3,
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "name",
          "email",
          "role",
          "is_active"
        ],
        "title": "UserUpdate",
        "description": "Esquema de entrada utilizado en PUT /users/{user_id}.\n\nA diferencia de UserBase, aquí `role` e `is_active` NO tienen valor por\ndefecto: los cuatro campos son obligatorios porque un PUT reemplaza por\ncompleto al usuario existente (actualización total).",
        "example": {
          "email": "usuario@ejemplo.com",
          "is_active": true,
          "name": "Usuario Ejemplo",
          "role": "admin"
        }
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}